migrated index to new style
