+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import {graphql} from 'gatsby';
 import * as React from 'react';
 
-import {FeaturedTile, PostTile, Button} from '../components';
+import {FeaturedTile, PostTile, Button, Link} from '../components';
 import {Layout} from '../layouts';
 import {Routes, useSiteData} from '../utils';
 
@@ -33,9 +33,18 @@ export default ({data}: IndexPageProps) => {
       <div className="my-huge">{titleMarkup}</div>
       <div className="space-y-big">
         <FeaturedTile node={featuredPost} />
-        {posts.map((post: Mdx) => (
-          <PostTile node={post} key={post.slug!} />
-        ))}
+        <div className="space-y-normal">
+          {posts.map((post: Mdx) => (
+            <div key={post.slug!}>
+              <div>
+                <Link url={Routes.blogPost(post.slug!)}>{post.title}</Link>
+              </div>
+              <div className="text-text-muted dark:text-text-muted-dark">
+                {post.date}
+              </div>
+            </div>
+          ))}
+        </div>
       </div>
       <Button className="mt-big mb-large" url={Routes.blog()}>
         Read more posts &rarr;
@@ -61,7 +70,7 @@ export const query = graphql`
         image {
           source {
             childImageSharp {
-              gatsbyImageData(width: 800)
+              gatsbyImageData(height: 420)
             }
           }
           alt
