+++ src/styles/global.scss
@@ -37,6 +37,7 @@ html {
 
 body {
   @apply font-sans;
+  @apply transition-colors;
 
   @media (prefers-color-scheme: light) {
     background-color: theme('colors.body-bg');
